ugc_api/pyproject.toml
[tool.poetry]
name = "ugc_api"
version = "0.1.0"
description = ""
authors = ["Pavel Petrikovskii <petrikovskii@hu.bpcbt.com>"]

[tool.poetry.dependencies]
python = "^3.7"
fastapi = "^0.78.0"
uvicorn = "^0.17.6"
aiokafka = "^0.7.2"
PyJWT = "^2.4.0"
sentry-sdk = "^1.5.12"

[tool.poetry.dev-dependencies]
black = "^22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
